tests/wpt/web-platform-tests/css/css-writing-modes/parsing/text-combine-upright-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes: getComputedValue().textCombineUpright</title>
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-text-combine-upright">
<meta name="assert" content="text-combine-upright computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-combine-upright", "none");
test_computed_value("text-combine-upright", "all");
</script>
</body>
</html>
